+++ ICD.Connect.Settings/Attributes/SettingsProperties/ControlPortParentSettingsPropertyAttribute.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Common.Properties;
+
+namespace ICD.Connect.Settings.Attributes.SettingsProperties
+{
+	[MeansImplicitUse]
+	public sealed class ControlPortParentSettingsPropertyAttribute : AbstractSettingsPropertyAttribute
+	{
+	}
+}