+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Added ControlPortParentSettingsPropertyAttribute for better parent device binding on DeployAV
+
 ### Changed
  - Fail gracefully when two settings attributes share the same factory name